Move TransportPool to its own package The experimental managed transport can also leverage TransportPool, moving it to its own package to accommodate that use case.
